Split into lib and binary
