+++ twackup-cli/Cargo.toml
@@ -1,0 +1,65 @@
+[package]
+name = "twackup-cli"
+version = "1.0.0"
+description = "Safe and fast iOS CLI utility for rebuilding DEBs and backing up packages"
+authors = ["DanP"]
+edition = "2021"
+publish = false
+license = "GPL-3.0-only"
+
+[[bin]]
+name = "twackup"
+path = "src/main.rs"
+
+[package.metadata.deb_pkg]
+name = "Twackup (CLI)"
+identifier = "ru.danpashin.twackup"
+section = "Utilities"
+priority = "optional"
+architecture = "iphoneos-arm"
+depends = "firmware (>= 3.0)"
+changelog = "changelog"
+homepage = ""
+assets = [
+    ["build/twackup-ios", "/usr/bin/twackup"],
+    ["COPYING", "/usr/share/doc/ru.danpashin.twackup/LICENSE"],
+]
+license_type = "GPL-3.0"
+
+[dependencies]
+twackup = { path = "../twackup", features = ["cli"] }
+
+# Core
+async-trait = "0.1"
+chrono = { version = "0.4", default-features = false, features = ["clock"] }
+tokio = { version = "1.5", features = ["rt-multi-thread", "sync", "macros"] }
+futures = "0.3"
+
+# Archiving
+flate2 = { version = "1.0", default-features = false, features = ["zlib"] }
+
+# Serializing
+plist = { version = "1.0", default-features = false }
+serde = { version = "1.0", features = ["derive"] }
+serde_any = { version = "0.5", default-features = false, features = ["yaml", "toml", "json"] }
+
+# Logging
+log = "0.4"
+
+# CLI Utils
+ansi_term = { version = "0.12", default-features = false, features = [] }
+indicatif = { version = "0.17", default-features = false, features = [] }
+stderrlog = { version = "0.5", default-features = false, features = [] }
+
+# Misc
+libc = "0.2"
+gethostname = "0.4"
+
+[dependencies.clap]
+version = "4.0"
+default-features = false
+features = ["derive", "std", "help", "usage", "error-context"]
+
+[features]
+default = ["ios"]
+ios = []
