+++ twackup-cli/src/commands/backup/import.rs
@@ -18,15 +18,15 @@
  */
 
 use super::{DataFormat, DataLayout, RepoGroup, RepoGroupFormat};
-use crate::{
-    cli::{commands::CliCommand, context::Context, ROOT_WARN_MESSAGE},
-    error::{Error, Result},
-    process,
-};
+use crate::{commands::CliCommand, context::Context, ROOT_WARN_MESSAGE};
 use std::{
     fs::File,
     io::{self, BufWriter, Write},
     process::{Command, Stdio},
+};
+use twackup::{
+    error::{Error, Result},
+    process,
 };
 
 #[derive(clap::Parser)]
