+++ twackup-cli/src/commands/build.rs
@@ -18,16 +18,7 @@
  */
 
 use super::CliCommand;
-use crate::{
-    builder::{
-        deb::{DebTarArchive, TarArchive},
-        Preferences, Worker,
-    },
-    cli::{context::Context, ADMIN_DIR, ROOT_WARN_MESSAGE, TARGET_DIR},
-    error::Result,
-    package::*,
-    progress::Progress,
-};
+use crate::{context::Context, ADMIN_DIR, ROOT_WARN_MESSAGE, TARGET_DIR};
 use chrono::Local;
 use gethostname::gethostname;
 use std::{
@@ -36,6 +27,15 @@ use std::{
     iter::Iterator,
     path::PathBuf,
     sync::{Arc, Mutex},
+};
+use twackup::{
+    builder::{
+        deb::{DebTarArchive, TarArchive},
+        Preferences, Worker,
+    },
+    error::Result,
+    package::*,
+    progress::Progress,
 };
 
 const DEFAULT_ARCHIVE_NAME: &str = "%host%_%date%.tar.gz";
