+++ twackup-cli/src/commands/leaves.rs
@@ -18,12 +18,12 @@
  */
 
 use super::{CliCommand, Context};
-use crate::{cli::ADMIN_DIR, error::Result};
+use crate::ADMIN_DIR;
 use std::path::PathBuf;
+use twackup::error::Result;
 
 #[derive(clap::Parser)]
-#[clap(version)]
-pub(crate) struct List {
+pub(crate) struct Leaves {
     /// Use custom dpkg <directory>.
     /// This option is used for detecting installed packages
     #[clap(long, default_value = ADMIN_DIR, value_parser)]
@@ -31,19 +31,13 @@ pub(crate) struct List {
 }
 
 #[async_trait::async_trait]
-impl CliCommand for List {
+impl CliCommand for Leaves {
     async fn run(&self, context: Context) -> Result<()> {
-        let packages = context.packages(&self.admindir, false).await?;
+        let packages = context.packages(&self.admindir, true).await?;
 
-        for (position, (_, package)) in packages.into_iter().enumerate() {
+        for (_, package) in packages {
             let section_sym = package.section.color().paint("▶︎");
-            println!(
-                "{:3}: {} {} - {}",
-                position + 1,
-                section_sym,
-                package.name,
-                package.id
-            );
+            println!("{} {} - {}", section_sym, package.name, package.id);
         }
 
         Ok(())
