+++ twackup-cli/src/commands/mod.rs
@@ -25,7 +25,7 @@ mod list;
 mod backup;
 
 use super::Context;
-use crate::error::Result;
+use twackup::error::Result;
 
 #[async_trait::async_trait]
 pub(crate) trait CliCommand {
