+++ twackup/Cargo.toml
@@ -1,0 +1,41 @@
+[package]
+name = "twackup"
+version = "1.0.0"
+description = "Safe and fast iOS CLI utility for rebuilding DEBs and backing up packages"
+authors = ["DanP"]
+edition = "2021"
+publish = false
+license = "GPL-3.0-only"
+
+[lib]
+crate-type = ["staticlib", "rlib"]
+
+[dependencies]
+
+# Core
+tokio = { version = "1.5", features = ["rt-multi-thread", "sync", "macros"] }
+
+# Archiving
+ar = "0.9"
+flate2 = { version = "1.0", default-features = false, features = ["zlib"] }
+tar = { version = "0.4", default-features = false }
+
+# Serializing
+plist = { version = "1.0", default-features = false }
+serde = { version = "1.0", features = ["derive"] }
+
+# Logging
+log = "0.4"
+thiserror = "1.0"
+
+# Misc
+libc = "0.2"
+libproc = { git = "https://github.com/danpashin/libproc-rs", branch = "0.12.0-ios", default-features = false }
+memmap = "0.7"
+
+# CLI Utils
+ansi_term = { version = "0.12", default-features = false, features = [], optional = true }
+
+[features]
+default = []
+cli = ["dep:ansi_term"]
